update : Checked the water Track Screen and rechecked all the Track Screens because the api is being changed

+++ mobile/src/main/java/fit/asta/health/navigation/track/model/net/water/AmountConsumed.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class AmountConsumed(
     @SerializedName("dailyAvg")
-    val dailyAvg: Double,
+    val dailyAvg: Float,
     @SerializedName("totalAmt")
-    val totalAmt: Int
+    val totalAmt: Float
 )
